@@ -1,79 +1,76 @@
 {
-  "name": "laravel/laravel",
-  "type": "project",
-  "description": "The skeleton application for the Laravel framework.",
-  "keywords": [
-    "laravel",
-    "framework"
-  ],
-  "license": "MIT",
-  "require": {
-    "php": "^8.2",
-    "aws/aws-sdk-php": "^3.262",
-    "binarytorch/larecipe": "^2.7",
-    "guzzlehttp/guzzle": "^7.2",
-    "inertiajs/inertia-laravel": "^1.0",
-    "laravel/framework": "^11.0",
-    "laravel/sanctum": "^4.0",
-    "laravel/tinker": "^2.8",
-    "tightenco/ziggy": "^1.0"
-  },
-  "require-dev": {
-    "fakerphp/faker": "^1.9.1",
-    "laravel/breeze": "^2.0",
-    "laravel/envoy": "^2.9",
-    "laravel/pint": "^1.17",
-    "laravel/sail": "^1.22",
-    "mockery/mockery": "^1.4.4",
-    "nunomaduro/collision": "^8.1",
-    "phpunit/phpunit": "^10.1",
-    "spatie/laravel-ignition": "^2.0"
-  },
-  "autoload": {
-    "psr-4": {
-      "App\\": "app/",
-      "Database\\Factories\\": "database/factories/",
-      "Database\\Seeders\\": "database/seeders/"
+    "name": "laravel/laravel",
+    "type": "project",
+    "description": "The skeleton application for the Laravel framework.",
+    "keywords": ["laravel", "framework"],
+    "license": "MIT",
+    "require": {
+        "php": "^8.2",
+        "aws/aws-sdk-php": "^3.262",
+        "binarytorch/larecipe": "v2.8.0",
+        "guzzlehttp/guzzle": "^7.2",
+        "inertiajs/inertia-laravel": "^v2.0.2",
+        "laravel/framework": "^12.0",
+        "laravel/sanctum": "^4.0",
+        "laravel/tinker": "^2.8",
+        "tightenco/ziggy": "^1.0"
     },
-    "files": [
-      "app/Helpers/globals.php"
-    ]
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "Tests\\": "tests/"
-    }
-  },
-  "scripts": {
-    "helper": "php artisan optimize:clear && php artisan ide-helper:generate && php artisan ide-helper:models -M && php artisan ide-helper:meta",
-    "post-autoload-dump": [
-      "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
-      "@php artisan package:discover --ansi"
-    ],
-    "post-update-cmd": [
-      "@php artisan vendor:publish --tag=laravel-assets --ansi --force"
-    ],
-    "post-root-package-install": [
-      "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
-    ],
-    "post-create-project-cmd": [
-      "@php artisan key:generate --ansi"
-    ]
-  },
-  "extra": {
-    "laravel": {
-      "dont-discover": []
-    }
-  },
-  "config": {
-    "optimize-autoloader": true,
-    "preferred-install": "dist",
-    "sort-packages": true,
-    "allow-plugins": {
-      "pestphp/pest-plugin": true,
-      "php-http/discovery": true
-    }
-  },
-  "minimum-stability": "stable",
-  "prefer-stable": true
+    "require-dev": {
+        "fakerphp/faker": "^1.9.1",
+        "laravel/breeze": "^2.0",
+        "laravel/envoy": "^2.9",
+        "laravel/pint": "^1.17",
+        "laravel/sail": "^1.22",
+        "mockery/mockery": "^1.4.4",
+        "nunomaduro/collision": "^8.1",
+        "phpunit/phpunit": "^11.0",
+        "spatie/laravel-ignition": "^2.0"
+    },
+    "autoload": {
+        "psr-4": {
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        },
+        "files": [
+            "app/Helpers/globals.php"
+        ]
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Tests\\": "tests/"
+        }
+    },
+    "scripts": {
+        "helper": "php artisan optimize:clear && php artisan ide-helper:generate && php artisan ide-helper:models -M && php artisan ide-helper:meta",
+        "post-autoload-dump": [
+            "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
+            "@php artisan package:discover --ansi"
+        ],
+        "post-update-cmd": [
+            "@php artisan vendor:publish --tag=laravel-assets --ansi --force"
+        ],
+        "post-root-package-install": [
+            "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
+        ],
+        "post-create-project-cmd": [
+            "@php artisan key:generate --ansi"
+        ]
+    },
+    "extra": {
+        "laravel": {
+            "dont-discover": []
+        }
+    },
+    "config": {
+        "optimize-autoloader": true,
+        "preferred-install": "dist",
+        "sort-packages": true,
+        "allow-plugins": {
+            "pestphp/pest-plugin": true,
+            "php-http/discovery": true
+        }
+    },
+    "minimum-stability": "stable",
+    "prefer-stable": true
 }
