@@ -3,6 +3,7 @@
 namespace Tests\Unit\Emails;
 
 use App\Mail\ContactEmail;
+use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 
 class ContactEmailTest extends TestCase
@@ -15,15 +16,13 @@ class ContactEmailTest extends TestCase
         $this->email = new ContactEmail('some@mail.com', 'some message');
     }
 
-    /** @test */
-    public function it_can_create_a_contact_email(): void
+    #[Test] public function it_can_create_a_contact_email(): void
     {
         $this->assertEquals('some@mail.com', $this->email->email);
         $this->assertEquals('some message', $this->email->message);
     }
 
-    /** @test */
-    public function it_can_return_expected_content()
+    #[Test] public function it_can_return_expected_content()
     {
         $content = $this->email->content();
 
