@@ -1,37 +1,40 @@
-import { defineConfig } from 'vite';
-import laravel from 'laravel-vite-plugin';
-import vue from '@vitejs/plugin-vue';
-import i18n from 'laravel-vue-i18n/vite';
-import path from 'path';
+import { defineConfig } from 'vite'
+import laravel from 'laravel-vite-plugin'
+import vue from '@vitejs/plugin-vue'
+import i18n from 'laravel-vue-i18n/vite'
+import path from 'path'
+import tailwindcss from '@tailwindcss/vite'
 
 export default defineConfig({
-    server: {
-        hmr: {
-            host: 'localhost'
+  server: {
+    // host: 'localhost',
+    hmr: {
+      host: 'localhost',
+    },
+  },
+  plugins: [
+    tailwindcss(),
+    laravel({
+      input: 'resources/js/app.js',
+      refresh: true,
+    }),
+    vue({
+      template: {
+        transformAssetUrls: {
+          base: null,
+          includeAbsolute: false,
         },
+      },
+    }),
+    i18n(),
+  ],
+  commonjsOptions: {
+    esmExternals: true,
+  },
+  resolve: {
+    alias: {
+      // eslint-disable-next-line no-undef
+      '@': path.resolve(__dirname, 'resources/js'),
     },
-    plugins: [
-        laravel({
-            input: 'resources/js/app.js',
-            refresh: true,
-        }),
-        vue({
-            template: {
-                transformAssetUrls: {
-                    base: null,
-                    includeAbsolute: false,
-                },
-            },
-        }),
-        i18n(),
-    ],
-    commonjsOptions: {
-        esmExternals: true
-    },
-    resolve: {
-        alias: {
-            // eslint-disable-next-line no-undef
-            '@': path.resolve(__dirname, 'resources/js'),
-        },
-    },
-});
+  },
+})
